pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ilo_auto"
version = "666.0"
description = "Generate HW healthcheck report from ILOM"
authors = [{name = "Shihab Istiak Saikat", email = "saikat.pro1990@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
keywords = ["health check", "ilo health check", "healthcheck"]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

dependencies = ["python-hpilo>=4.4.3", "pandas>=2.2.3", "openpyxl>=3.1.5"]


[tool.setuptools]
packages = ["ilo_auto"]

[project.scripts]
hcilo = "ilo_auto.hcilo:main"